packages/gb_regressor/requirements_dev.txt
-r requirements.txt

# testing requirements
pytest==7.1.3

# linear regression model solving the same problem as our GB regressor
# this is for differential testing purposes
tid-regression-model==3.1.2